in App.js
Hello! 
I was reading your article on Medium. Thanks.
I was write my versión in Spanish for me (with your credit).
I'm a beginner and it confused me the line 37 in App.js, so I said to my self: ¿where is the 'VideoModel' in the stack?
I don't know if I'm wrong.
Thanks for the article and this repository.